MAINT: rename cymru parsers to parser_*
